Avoid extra redirect with spee.ch images
I'm not sure what's the history behind splitting the logic between speech and non-speech, but currently, doing

`${thumbnail}?quality=${quality}&height=${height}&width=${width}`

will result in a redirect to

`${THUMBNAIL_CDN_URL}s:${width}:${height}/quality:80/plain/${thumbnail}` (hardcoded to 80).

That's the same CDN url we use for non-speech, so might as do the same for all images to avoid the redirect.